refactor: header, container 리팩토링
