+++ app/components/Layout/Header/style.tsx
@@ -1,3 +1,4 @@
+import { DefaultInnerContainer } from 'components/Container/style';
 import styled from 'styled-components';
 
 export const HeaderWrapper = styled.header`
@@ -63,7 +64,13 @@ export const HeaderWrapper = styled.header`
   }
 `;
 
-export const LogoContainer = styled.div`
+export const InnerContainer = styled(DefaultInnerContainer)`
+  justify-content: space-between;
+  align-items: center;
+  padding: 20px 0;
+`;
+
+export const MainLogoContainer = styled.div`
   display: flex;
   align-items: center;
